# openapi_schema.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "OIDC Authentication Server",
    "description": "OpenID Connect compliant authentication and authorization server",
    "version": "1.0.0",
    "contact": {
      "name": "OIDC Platform Support",
      "url": "https://github.com/your-org/oidc-platform"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/login": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Login Page",
        "description": "Display the login page",
        "operationId": "login_page",
        "parameters": [
          {
            "name": "redirect_uri",
            "in": "query",
            "description": "URL to redirect after successful login",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login page HTML",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Authentication"],
        "summary": "User Login",
        "description": "Authenticate user credentials",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirect on successful login"
          },
          "400": {
            "description": "Invalid credentials",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Registration Page",
        "description": "Display the user registration page",
        "operationId": "register_page",
        "responses": {
          "200": {
            "description": "Registration page HTML",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Authentication"],
        "summary": "User Registration",
        "description": "Register a new user",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirect to login on successful registration"
          },
          "400": {
            "description": "Registration failed",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/oauth2/authorize": {
      "get": {
        "tags": ["OAuth2/OIDC"],
        "summary": "Authorization Endpoint",
        "description": "OAuth 2.0 authorization endpoint for Authorization Code Flow",
        "operationId": "authorization_endpoint",
        "parameters": [
          {
            "name": "response_type",
            "in": "query",
            "required": true,
            "description": "OAuth 2.0 response type",
            "schema": {
              "type": "string",
              "enum": ["code"]
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "description": "OAuth 2.0 client identifier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": true,
            "description": "Client redirect URI",
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": true,
            "description": "OAuth 2.0 scope values",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "description": "CSRF protection value",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code_challenge",
            "in": "query",
            "required": false,
            "description": "PKCE code challenge",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code_challenge_method",
            "in": "query",
            "required": false,
            "description": "PKCE code challenge method",
            "schema": {
              "type": "string",
              "enum": ["S256", "plain"]
            }
          },
          {
            "name": "nonce",
            "in": "query",
            "required": false,
            "description": "OpenID Connect nonce",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consent page HTML",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to login if not authenticated"
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/consent": {
      "post": {
        "tags": ["OAuth2/OIDC"],
        "summary": "User Consent",
        "description": "Handle user consent for authorization",
        "operationId": "consent_endpoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirect to client with authorization code or error"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/oauth2/token": {
      "post": {
        "tags": ["OAuth2/OIDC"],
        "summary": "Token Endpoint",
        "description": "Exchange authorization code for tokens",
        "operationId": "token_endpoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuth2Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid client credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuth2Error"
                }
              }
            }
          }
        }
      }
    },
    "/oauth2/userinfo": {
      "get": {
        "tags": ["OAuth2/OIDC"],
        "summary": "UserInfo Endpoint",
        "description": "Get user information using access token",
        "operationId": "userinfo_endpoint",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": ["email", "password"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "confirm_password": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          }
        },
        "required": ["email", "password", "confirm_password", "first_name", "last_name"]
      },
      "ConsentRequest": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": ["allow", "deny"]
          },
          "response_type": {
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string",
            "format": "uri"
          },
          "scope": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "code_challenge": {
            "type": "string"
          },
          "code_challenge_method": {
            "type": "string"
          },
          "nonce": {
            "type": "string"
          }
        },
        "required": ["action", "response_type", "client_id", "redirect_uri", "scope"]
      },
      "TokenRequest": {
        "type": "object",
        "properties": {
          "grant_type": {
            "type": "string",
            "enum": ["authorization_code"]
          },
          "code": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string",
            "format": "uri"
          },
          "client_id": {
            "type": "string"
          },
          "client_secret": {
            "type": "string"
          },
          "code_verifier": {
            "type": "string"
          }
        },
        "required": ["grant_type", "code", "redirect_uri", "client_id"]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "OAuth 2.0 access token"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "description": "Token lifetime in seconds"
          },
          "refresh_token": {
            "type": "string",
            "description": "OAuth 2.0 refresh token"
          },
          "scope": {
            "type": "string",
            "description": "Granted scopes"
          },
          "id_token": {
            "type": "string",
            "description": "OpenID Connect ID token"
          }
        },
        "required": ["access_token", "token_type"]
      },
      "UserInfo": {
        "type": "object",
        "properties": {
          "sub": {
            "type": "string",
            "description": "Subject identifier"
          },
          "name": {
            "type": "string",
            "description": "Full name"
          },
          "given_name": {
            "type": "string",
            "description": "Given name"
          },
          "family_name": {
            "type": "string",
            "description": "Family name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address"
          },
          "email_verified": {
            "type": "boolean",
            "description": "Email verification status"
          }
        },
        "required": ["sub"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["detail"]
      },
      "OAuth2Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "OAuth 2.0 error code"
          },
          "error_description": {
            "type": "string",
            "description": "Human-readable error description"
          },
          "error_uri": {
            "type": "string",
            "format": "uri",
            "description": "URI to error information page"
          }
        },
        "required": ["error"]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OAuth 2.0 Bearer token authentication"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "user_id",
        "description": "Session-based authentication using HTTP-only cookies"
      }
    }
  },
...
}